pyproject.toml
[project]
name = "mxwhisper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.119.0",
    "gtts>=2.5.4",
    "httpx>=0.28.1",
    "openai-whisper>=20250625",
    "pydantic-settings>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
    "python-json-logger>=2.0.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "temporalio>=1.18.1",
    "uvicorn>=0.37.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*", "docker*", "config*", "alembic*", "instructions*"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=0.25.0",
    "ruff>=0.14.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E402", "E722"]  # Allow import not at top, bare except for now

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[project.scripts]
mxwhisper = "app.cli:main"
mxwhisper-worker = "app.workflows.transcribe.worker:main"
